Significant Improvements in the word-filter function!
Previously, it was an array of regexes, and a for loop checking
whether or not the given word matches any of the regex.
Now, the array is an ROT1 encoded array (and the assignment of
words is divided into lines depending upon the type), and, inst-
ead of checking the given word with every single word in inapprop
array, it now greps whether or not the ROT1 encoded version of the
given word is in the inapprop_rot1 array.

Took a bit of time to do this. But this one's quite efficient, too.
And I'm quite happy about this. :)